Add variables.php
variables.php is a page we are going to create to save our most
requested variables from the database like $name , $money , $rank etc
etc. We are making safe so that we don't have to keep creating mysql
query on every page to connect to users and get us information.